+++ src/modules/user/user.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common';
-import { UserService } from 'src/services/user.service';
+import { UserService } from 'src/modules/user/user.service';
 
 @Controller('users')
 export class UserController {
