+++ src/modules/user/user.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { UserController } from './user.controller';
-import { UserService } from 'src/services/user.service';
+import { UserService } from 'src/modules/user/user.service';
 
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserEntity } from 'src/entities/user/user.entity';
+import { UserEntity } from 'src/shared/entities/user/user.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserEntity])],
