+++ src/shared/entities/user/user.entity.ts
